Add a method to easily add regression tests based on an input and expected output file (#218)
* Add a method to easily add unit/regression tests

The added test will perform all tests defined in the `valid-inputs`
directory. A correct test is defined by a folder that contains an input
file `input.in` and an expected output file `output.json`. The test will
run a simulation based on `input.in`, and compare its output to
`output.json`. For now this is a simple equivalency check on the resulting
JSON output object and the expected JSON object defined in `output.json`.

* Add valid input test case

* Add function to regenerate expected `output.json`

* Add a custom `json_equal` function

This function recursively compares two JSON objects. Notably
the user can supply a relative and absolute tolerance for
comparing floating point numbers.

* Document the new functions

* Add the `--regenerate` flag to `test_valid_inputs`

This flag allows for the user to regenerate the outputs of the
available regression tests. The user can also provide the name
of a specific subtest for which the output needs to be
regenerated.

* Regenerate output for `default-input-N2`