Throw when duplicate elastic cross sections are found
This error occurs when a parent and child state both carry an elastic
cross section. In that case the elastic contribution of the child
will be counted twice when computing the total elastic cross section
(assuming that its contribution is also captured by the elastic
cross section of the parent).