Fix and accept the new `findStatesToUpdate` implementation (#187)
* Do not throw when a state is not consumed in reactions

Previously this behavior was desired, as every electronic state with
non-zero population should have at least an elastic cross section.
This is no longer the case however, as it is also valid for the state
to have an elastic cross section for all its populated (grand)children.

* Remove the `NEW_FINDSTATESTOUPDATE_IMPLEMENTATION` flag

The old implementation has been removed.

* Throw when duplicate elastic cross sections are found

This error occurs when a parent and child state both carry an elastic
cross section. In that case the elastic contribution of the child
will be counted twice when computing the total elastic cross section
(assuming that its contribution is also captured by the elastic
cross section of the parent).

* Add test for duplicate elastic cross sections

Test that LoKI-B correctly throws an error when both the parent
and populated child state carry an elastic cross section.

* Document the `findStatesToUpdate` function

* Refactor `findStatesToUpdate`

* Trim the header of the LXCat test inputs

* Add test for throwing when an elastic/effective cs is missing