Use the serialized state description to parse LXCat 3 species definitions (#219)
* Initial efforts to support current version of schema

* Use the first `info` object and add `FIXME`s

* Add incomplete parsing of states

Uses the `serialized` property of state definitions. However,
this is only available on data downloaded from the new LXCat.
How do we treat local files that do not contain serialized
versions of the state definitions.

* Correctly grab the gas name

Split the charge description from the gas name if available.

* Check for the electron by species type

* Grab the charge from the detailed state description

* Check that compound states are not empty

* Cleanup comments

* Parse integers from summary strings for compound vibrot quanta

* Parse charge from the composition summary

* Do not special case electrons

* Update LXCat JSON in Druyvesteyn test

* Update nonuniform inelastic delta test JSON inputs

* Update LXCat JSON files

* Add unit tests for JSON state parsing

* Correctly throw when the user provides a compound for an unsupported species

LoKI-B currently only supports compounds for states with a single quanta on the
compound level.

* Update inputs to include the species type

* Add tests to check whether LoKI throws for unsupported compounds

We currently only support vibrational and rotational compound states
for species types that have a single quantum on the compound level.

* Support compound electronic states in JSON input

* Support vibrational compounds for states types beyond diatoms

* Fix loading of effective populations

Use the string summary to find the specified state instead
of its id (which is unreliable). This is similar to how
loading of state properties is implemented.

This patch also implements wildcard states when loading
effective populations.

* Remove unused import

* Fix signed-unsigned comparison warnings

* Fix error message for compound rotational states

It erroneously mentioned vibrational compounds, which are nowk
supported.

* Update broken tests

Some tests were lagging behind recent changes in the
species parsing and had to be updated.