Throw proper error on unknown species when loading state property (#231)
* Throw proper error on unknown species when loading state property

Previously the call to the `std::runtime_error` constructor would try to
access the `state` key on the `propEntry` JSON object. This key does not
generally exist on this object, which would in turn throw a different
error. This lead to unclear feedback to the user. The problem is fixed
by using the available `state_id` instead.

Resolves #211

* Use `Log` instead of `std::cout`

This will reroute log messages to `std::cerr` such that log output
is not rerouted to `gnuplot` when piping.