Correctly serialize species charge (#230)
* Introduce the `serialize_charge` function

This function is used to serialize the charge integer that
is obtained from cross sections defined in JSON format.

* Serialize the charge conform the LoKI-B manual

The LoKI-B manual suggests that charge should be defined by repeating
`+` or `-` for positive and negative charge respectively. This means
that e.g. `3+` becomes `+++`, and `2-` becomes `--`. This does not
scale well to higher absolute charge numbers. Perhaps we should change
this behavior at some point.

* Remove trailing comma in serialized charge state

The comma is now only added when both a charge and an electronic description
are present.

* Unit test the charge serialization

* Complete Druyvesteyn test input file