Correctly detect charge states
The second capture group of the regular expression in `propertyStateFromString`
would only capture the charge of a given state when it also had an electronic
descriptor. Otherwise, the third capture group would capture the charge state,
LoKI-B would then erroneously load the state as an electronic state. This patch
fixes the regular expression such that the second capture group always (greedily)
matches the charge state (also adding support for multiply ionized states), and
it fixes the surrounding logic to correctly classify the state.