Fix charge state parsing and serialization (#242)
* Correctly detect charge states

The second capture group of the regular expression in `propertyStateFromString`
would only capture the charge of a given state when it also had an electronic
descriptor. Otherwise, the third capture group would capture the charge state,
LoKI-B would then erroneously load the state as an electronic state. This patch
fixes the regular expression such that the second capture group always (greedily)
matches the charge state (also adding support for multiply ionized states), and
it fixes the surrounding logic to correctly classify the state.

* Fix (charge) state serialization

Charge states now no longer expect a trailing comma in their
body, e.g. Ar(+,) is now simply Ar(+). Serialization has been
homogenized between the `StateEntry` and `Gas::State` classes.

* Cleanup `StateEntry` includes

* Fix error message

The wrong error was thrown when the population of the parent is 0, but
the total population of its children is >0.

* Distinguish serialized root and neutral charge states

Previously, both were serialized as `X()`. This patch serializes
root states simply as `X` instead.