Also iterate charge states when checking for elastic cross sections (#253)
* Iterate all states of the gas when checking for elastic cs

Previously, only the electronic, vibrational, and rotational states
were checked in `findStatesToUpdate`. This patch starts the check at
the root state instead. This means that a user can now apply an
elastic cross section to a charge state, or even a root state.

Resolves #252

* Only check for elastic cs from the charge state down

This means that an elastic cross section computed from an
effective cross section defined on one of the children of
`N2()` will be attached to `N2()`. Therefore it will not
be applied to other charge states such as `N2(+)`.

* Apply computed elastic cross section to electronic levels

This makes the behavior related to assigning elastic cross
sections computed from effective cross sections consistent
with the previous behavior.