Update mobility and spatial field operator discretization in spatial growth routine (#249)
* Add a function to integrate f_node*dg_cell/du

The function `fNodegPrimeEnergyIntegral` computes the discretized
derivative of the above expression for both uniform and nonuniform
grids.

* Fix discretization of mobility and spatial growth operator

Uses the newly introduced `fNodegPrimeIntegral` function to integrate
the mobility, directly using D0 defined on the nodes. Equivalently, the
spatial growth matrix discretization has been updated to use D0 defined
on the nodes as well.

One curiosity is that `D0` is not interpolated from its variant defined
on the nodes `D0Nodes`. Instead it is computed from the cell energies
and the total cell cross section (which is not equivalent). This means
that it cannot be used as a shorthand for `D0Nodes` values interpolated
to the cells (this caused some issues in the first implementation).

* Make integration of `powerMobility` consistent

In particular with the integration of the mobility and the discretization
of the spatial growth operator in the spatial growth routine.

* Remove superfluous faces to cells interpolation

* Make integration of output mobilities consistent

As far as I am aware the different mobilities are now computed
consistently throughout LoKI-B.

* Regenerate regression test output