Make spatial convergence log message conditional (#262)
* Make spatial convergence log message conditional

Previously, the user would be wrongly notified that the spatial growth
routine converged when the maximum number of iterations was exceeded.

Fixes #261

* Rename `hasConverged` -> `iterating`

The negation on the condition in the while loop has been
removed. The name `hasConverged` was a bit confusing, as
it was also set to true when the routine did not converge.