Add Id to collision representing input index (#206)
* Add Id to collision representing input index

When reading cross sections, a unique id is added to each
collision which represents the order in which the cross section
was read.
When the input contains an effective cross section, an additional
collision containing the derived elastic collision is created.
In that case, the id is -1.

The goal of this change is that a listener can identify an object
in EedfCollisionDataMixture.rateCoefficients() from its Id() member
linking it to an input process.

* Use lower case for (member) functions

* Copy the effective id over to the resulting elastic collisions

* Format using `clang-format`

---------

Co-authored-by: Wouter Graef <wgraef@epgmod.phys.tue.nl>
Co-authored-by: Daan Boer <daan-boer@hotmail.com>