Add the `InterpolatingIterator` class
This class allows to iterate over a vector y defined on domain x
for monotonically increasing values of x. The iterator will
provide correct value of the slope in y (linear or logarithmic)
for a given value x_current.

This class will form the basis for computing the analytical
collision integrals.

TODO: Implement boundary conditions.