Add different EEDF integrators
The three classes `ConstIntegrator`, `LinearIntegrator`, and
`LogIntegrator` provide stencils for the collision integral
matrix. A matrix set using one of the integrators multiplied
by the eedf results in the integrated value of the collision
integral.

These classes can be used to implement a collision integral
routine that build the matrices related to the inelastic
collision terms.